backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.40.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.8.0",
]
